backend/requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.10
pydantic-settings==2.10.1
requests==2.32.5
python-dateutil==2.9.0.post0
yt-dlp
TikTokApi
playwright
instaloader
